render the front page of the site
This was a complicated spec to fix. I worked on it over
the course of several days to untangle all the method calls.
(A few minutes at a time, not whole days.)

The take-home here is severalfold. First, excessive mocking
is definitely a smell indicating the code being tested is
poorly structured. Second, rspec mocks and stubs have changed
a fair bit between rspec 2 and rspec 3. Not being a big fan
of mocking in any case, it was annoying to deal with the
differences. However, my mocking-fu is improved as a result of
getting this spec working, so that's not entirely a bad thing.

Going forward from this commit, several things need to happen.
The spec files need to be renamed from erb to haml. The view needs
to be refactored in favor of using helpers which are much easier
to test, and the current set of mocks in the spec will get
reduced along the way.

It's interesting to see my Rails code from ~2010. Pretty terrible
stuff!