Chase down <title> spec failure
I'm almost surely testing title contents incorrectly, by
attempting to capture the behavior in the view. This is
because the content_for? call needs the definition of
`title` from the helper, which isn't called during view
specs. Not sure yet how I'm going to handle it, so this
commit fails a view spec for reminder.