projects controller spec progress
Cleaned up one of the pending update specs by deleting all
the code which was in the spec and rewriting it from scratch.
The other two update specs got their code deleted as well,
haven't yet written the updates for those.

Uncommented a delete spec to get handled in the near future.