project model spec - subject instead of before
More cleanup. Note that the model specs in this project
in part reflect its original test-first implementation.
For example, project belongs_to member, write a test for
the association, then implement the association. The result
of this test-first style is a large number of tests which
are essentially testing the Rails implementation as well.
Many of the current model tests could likely be removed.