profile show view spec fixed
More precisely, a pending spec was implemented. Apparently
this was nice TDD work that was unfinished: I had to add
the appropriate link to a view.

Next up is cleaning the controller specs from old stub and
mock techniques, and getting rid of a lot of the instance
variables in the controller and view specs.