rails51 - rails 5.1 has gem compatibility problems
There are at least two gem problems with rails 5.1:

1. cucumber-rails, which is the generator for cucumber
files. This is not really necessary for operating a rails
app and can be commented out for now.

2. devise has a conflict with railties, which is not yet
handled in a devise release at the time of this commit.
Hence the update won't work.

In the meantime, this commit takes rails to 5.1.0.rc2.
According the maintainers, devise should be updated
within a few days of this commit.