rails51 - remaining file additions for app:update
The remaining files added by the app:update rake task
are in this commit, notable `bin/`. The main thing left
with the app:update task is understanding how the new
default configurations work, but there is still a bug
in this part of the process, which may be fixed in 5.1.0
proper, which has to wait until devise updates.