cucumber: a very small, incremental advance on signin
I'm coming to the opinion that the whole signin feature
set of steps is completely borked. It might be easier
to redo them from scratch than to attempt to fix what
I have right now.

Redoing from scratch could be done as follows:

1. set all currently failing features to @wip.

2. copy the failing feature to a new feature/

3. add a bogon to every step of the new feature
to force cucumber to issue to steps.

4. Implement the new steps.

It would probably be helpful to get a better grip on
what devise is doing as well. I've forgotten so much.