migration-update - tag with [5.1]
For work reasons, I had to cycle my postgres installation,
hence needed to rebuild the database for the first time
since upgrading to rails 5.1. This was a bit of a hassle
with strong migrations gem installed, so I disabled that
to do the migration.

sed script from SO made the changes, very handy.